Added DataAccess Library And Updated The Shared Models
I added the dataaccess library and configured it to with with EntityFramework Core. The dataaccess for now has Movies, Reservations, Projections and CinemaRooms. This could be expanded in the future. The necessary CRUD operations have been implemented for those entities. Finally I made some minor changes in the shared models class.