Added Authentication And Authorization Library
I added an authentication and authorization library for the project. The user that is used is found in the shared library and it extends identity user. All necessary operations for easy authentication access have been implemented, but more can be added in the future. For the authorization the default roles are being used and all necessary operations for easy authorization have been implemented. There are also 3 seeded roles User, Content Admin and Admin.