Fully Implemented Sign Up, Sign In And Sign Out
I made the necessary to fully implement the sign in, sign up and sign up from the library to the front end, which means I added the views, the account controller and made some changes in the authentication and authorization library. I also added a sample homepage and a general view file structure of the app.