Added CinemaProgram
I added the cinema program view and controller. I still need to add the bankCards transaction for the tickets in that view, but I will have to add projections for that first, which will be done after the functionality for the content admin has been done and also when the bankCards functionality will be added for the users. I also changed a bit the layout and now the correct links in the navbar are being shown based on whether the user is signed in and their role.