Fully Implemented Admin
With this update the admin basic functionality(create user, edit user, delete user and manage users and filtering) has been fully implemented. That means that the authentication and authorization classes have been updated and the admin controller and the corresponding views have been added. I also changed all the disabled attributes in the editAccount view to the readonly attribute. Further changes will be implemented for the admin when the bankcards and tickets are added.