Fully Implemented Content Admin
With this update the content admin is fully implemented. That means that they can create update view and delete movies through the corresponding controller, controller actions and views that were added. Some minor modifications also were made in the dataaccess library for that purpose. The content admin with this update also have the ability to create projections for movies and delete projections from movies.