Implemented BankCard And Tickets
With this update the user is able to buy tickets and to refund tickets. The user is also able to add, edit and remove bankcards to their account. All the necessary changes have been made to the controllers, the models and the views for these purposes. This will be the last update, but a possible next update would be fixing a small bug were the user can buy tickets without adding a bankCard and also implementing the admin to be able to refund tickets and to manipulate the bankcards of the other accounts.